convert test_bt_customer_types
